+++ Museum/Exhibitor.cs
@@ -94,7 +94,7 @@ namespace Museum
             var values = new[] {Password, Name, Phone.ToString(), Mail};
             var insertPersons = SqlOperations.Instance.Insert(table, keys, values);
             Console.WriteLine(insertPersons);
-            Id = (int) DBConnection.Instance.Execute(insertPersons);
+            Id = DBConnection.Instance.Execute(insertPersons);
 
             table = "exhibitors";
             keys = new[] {TypeProperty, "persons_id"};
